+++ client/src/views/GameSettings.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import ErrorModal from './components/ErrorModal';
-import Connection from './workers/Conncetion';
+import ErrorModal from '../components/ErrorModal';
+import Connection from '../workers/Conncetion';
 import Behavior from './Behavior';
-import BehaviorsList from './BehaviorsList';
+import BehaviorsList from '../components/BehaviorsList';
 
-import { defaultBehavior } from './workers/Behaviors';
+import { defaultBehavior } from '../workers/Behaviors';
 
 const GameSettings = ({ }) => {
   // State for the game settings
