+++ client/src/views/Login.jsx
@@ -2,8 +2,8 @@
 
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom'; // Import useNavigate from react-router-dom
-import LoginForm from './components/LoginForm';
-import ErrorModal from './components/ErrorModal';  // Import the ErrorModal
+import LoginForm from '../containers/LoginForm';
+import ErrorModal from '../components/ErrorModal';  // Import the ErrorModal
 
 const Login = () => {
   const [email, setEmail] = useState('');
